+++ CineasteTests/WatchlistMovieCellTests.swift
@@ -10,8 +10,8 @@ import XCTest
 import CoreData
 @testable import Cineaste_App
 
-class MovieListCellTests: XCTestCase {
-    let cell = MovieListCell()
+class WatchlistMovieCellTests: XCTestCase {
+    let cell = WatchlistMovieCell()
 
     override func setUp() {
         super.setUp()
@@ -20,35 +20,36 @@ class MovieListCellTests: XCTestCase {
         cell.addSubview(poster)
         cell.poster = poster
 
-        let title = TitleLabel()
-        cell.addSubview(title)
-        cell.title = title
-
         let separatorView = UIView()
         cell.addSubview(separatorView)
         cell.separatorView = separatorView
 
-        let votes = DescriptionLabel()
-        cell.addSubview(votes)
-        cell.votes = votes
+        let releaseAndRuntime = UILabel()
+        cell.addSubview(releaseAndRuntime)
+        cell.releaseAndRuntimeLabel = releaseAndRuntime
 
-        let runtime = DescriptionLabel()
-        cell.addSubview(runtime)
-        cell.runtime = runtime
+        let title = UILabel()
+        cell.addSubview(title)
+        cell.title = title
 
-        let releaseDate = DescriptionLabel()
-        cell.addSubview(releaseDate)
-        cell.releaseDate = releaseDate
+        let voteView = VoteView()
+        cell.addSubview(voteView)
+        cell.voteView = voteView
     }
 
     func testConfigureShouldSetCellTitleAndVotesCorrectly() {
         cell.configure(with: storedMovie)
 
         XCTAssertEqual(cell.poster.image, UIImage.posterPlaceholder)
+        XCTAssertEqual(cell.releaseAndRuntimeLabel.text, storedMovie.formattedRelativeReleaseInformation
+            + " ∙ "
+            + storedMovie.formattedRuntime)
         XCTAssertEqual(cell.title.text, storedMovie.title)
-        XCTAssertEqual(cell.votes.text, storedMovie.formattedVoteAverage)
-        XCTAssertEqual(cell.runtime.text, storedMovie.formattedRuntime)
-        XCTAssertEqual(cell.releaseDate.text, storedMovie.formattedReleaseDate)
+
+        let nonbreakingSpace = "\u{00a0}"
+        XCTAssertEqual(cell.voteView.content,
+                       storedMovie.formattedVoteAverage
+                        + "\(nonbreakingSpace)/\(nonbreakingSpace)10")
     }
 
     private let storedMovie: StoredMovie = {